small issue
The code currently had an issue where it will only push one int into the vector<>. then the checkinlist() function would immediately see that its in the list and return true, which is wrong. So init the vector<> with the head of the list. and have the checkInList() function push onto the vector